Change result datatypes based on input type

+++ src/main/resources/com/nec/arrow/functions/cpp/transfer-definitions.h
@@ -33,7 +33,7 @@ typedef struct
 
 typedef struct
 {
-    long long *data;
+    long *data;
     int count;
 } non_null_bigint_vector;
